requirements.txt
aiohttp==0.21.6
chardet==2.3.0
discord.py==0.10.0
websockets==3.1
